- Restructured the folder layout for improved clarity. - Added support for passing a label prop to dynamically set the content inside the elatedProduct
