+++ app/page.js
@@ -1,13 +1,13 @@
 "use client";
 
 import Hero from "./hero/page";
-import RelatedProduct from "./relatedProduct";
+import RelatedProduct from "../common/relatedProduct";
 
 export default function Home() {
   return (
     <div className="md:flex flex-col flex gap-16 items-center justify-center ">
       <Hero />
-      <RelatedProduct />
+      <RelatedProduct heading="RelatedProduct" />
     </div>
   );
 }
